落下速度追加

+++ Object/mino.h
@@ -26,7 +26,7 @@ public:
 	int m_getPosX()const { return m_posX; }
 	int m_getPosY()const { return m_posY; }
 
-	int m_getCount()const { return m_countY; }
+	int m_getCount()const { return (int)m_countY; }
 
 private:
 	int m_block[BLOCK_HEIGHT][BLOCK_WIDTH];
